Extract payload type
