+++ protocol/flows/src/ctx.rs
@@ -1,8 +1,9 @@
 use async_trait::async_trait;
 use consensus_core::api::DynConsensus;
 use kaspa_core::info;
-use p2p_lib::core::{ConnectionError, ConnectionInitializer, KaspadMessagePayloadType, Router};
+use p2p_lib::core::{ConnectionError, ConnectionInitializer, Router};
 use p2p_lib::pb::{self, kaspad_message::Payload, KaspadMessage, VersionMessage};
+use p2p_lib::KaspadMessagePayloadType;
 use std::sync::Arc;
 use tokio::sync::mpsc::Receiver;
 use uuid::Uuid;
