+++ protocol/p2p/src/core.rs
@@ -1,7 +1,7 @@
 use crate::pb::{
-    kaspad_message::Payload as KaspadMessagePayload, p2p_client::P2pClient as ProtoP2pClient, p2p_server::P2p as ProtoP2p,
-    p2p_server::P2pServer as ProtoP2pServer, KaspadMessage,
+    p2p_client::P2pClient as ProtoP2pClient, p2p_server::P2p as ProtoP2p, p2p_server::P2pServer as ProtoP2pServer, KaspadMessage,
 };
+use crate::KaspadMessagePayloadType;
 use futures::FutureExt;
 use kaspa_core::{debug, error, info, trace, warn};
 use std::net::ToSocketAddrs;
@@ -17,54 +17,6 @@ use tokio_stream::StreamExt;
 use tonic::transport::{Error as TonicError, Server as TonicServer};
 use tonic::{Request, Response, Status as TonicStatus, Streaming};
 use uuid::Uuid;
-
-#[repr(u8)]
-#[derive(Debug, Copy, Clone)]
-pub enum KaspadMessagePayloadType {
-    Addresses = 0,
-    Block,
-    Transaction,
-    BlockLocator,
-    RequestAddresses,
-    RequestRelayBlocks,
-    RequestTransactions,
-    IbdBlock,
-    InvRelayBlock,
-    InvTransactions,
-    Ping,
-    Pong,
-    Verack,
-    Version,
-    TransactionNotFound,
-    Reject,
-    PruningPointUtxoSetChunk,
-    RequestIbdBlocks,
-    UnexpectedPruningPoint,
-    IbdBlockLocator,
-    IbdBlockLocatorHighestHash,
-    RequestNextPruningPointUtxoSetChunk,
-    DonePruningPointUtxoSetChunks,
-    IbdBlockLocatorHighestHashNotFound,
-    BlockWithTrustedData,
-    DoneBlocksWithTrustedData,
-    RequestPruningPointAndItsAnticone,
-    BlockHeaders,
-    RequestNextHeaders,
-    DoneHeaders,
-    RequestPruningPointUtxoSet,
-    RequestHeaders,
-    RequestBlockLocator,
-    PruningPoints,
-    RequestPruningPointProof,
-    PruningPointProof,
-    Ready,
-    BlockWithTrustedDataV4,
-    TrustedData,
-    RequestIbdChainBlockLocator,
-    IbdChainBlockLocator,
-    RequestAnticone,
-    RequestNextPruningPointAndItsAnticoneBlocks,
-}
 
 #[derive(Debug)]
 pub(crate) enum HubEvent {
@@ -341,7 +293,7 @@ pub struct Router {
     identity: Uuid,
 
     /// Routing map for mapping messages to subscribed flows
-    routing_map: RwLock<HashMap<u8, MpscSender<KaspadMessage>>>,
+    routing_map: RwLock<HashMap<KaspadMessagePayloadType, MpscSender<KaspadMessage>>>,
 
     /// The outgoing route for sending messages to this peer
     outgoing_route: MpscSender<KaspadMessage>,
@@ -412,61 +364,6 @@ impl Router {
         self.identity
     }
 
-    fn payload_to_u8(payload: &KaspadMessagePayload) -> u8 {
-        let res = match payload {
-            KaspadMessagePayload::Addresses(_) => KaspadMessagePayloadType::Addresses,
-            KaspadMessagePayload::Block(_) => KaspadMessagePayloadType::Block,
-            KaspadMessagePayload::Transaction(_) => KaspadMessagePayloadType::Transaction,
-            KaspadMessagePayload::BlockLocator(_) => KaspadMessagePayloadType::BlockLocator,
-            KaspadMessagePayload::RequestAddresses(_) => KaspadMessagePayloadType::RequestAddresses,
-            KaspadMessagePayload::RequestRelayBlocks(_) => KaspadMessagePayloadType::RequestRelayBlocks,
-            KaspadMessagePayload::RequestTransactions(_) => KaspadMessagePayloadType::RequestTransactions,
-            KaspadMessagePayload::IbdBlock(_) => KaspadMessagePayloadType::IbdBlock,
-            KaspadMessagePayload::InvRelayBlock(_) => KaspadMessagePayloadType::InvRelayBlock,
-            KaspadMessagePayload::InvTransactions(_) => KaspadMessagePayloadType::InvTransactions,
-            KaspadMessagePayload::Ping(_) => KaspadMessagePayloadType::Ping,
-            KaspadMessagePayload::Pong(_) => KaspadMessagePayloadType::Pong,
-            KaspadMessagePayload::Verack(_) => KaspadMessagePayloadType::Verack,
-            KaspadMessagePayload::Version(_) => KaspadMessagePayloadType::Version,
-            KaspadMessagePayload::TransactionNotFound(_) => KaspadMessagePayloadType::TransactionNotFound,
-            KaspadMessagePayload::Reject(_) => KaspadMessagePayloadType::Reject,
-            KaspadMessagePayload::PruningPointUtxoSetChunk(_) => KaspadMessagePayloadType::PruningPointUtxoSetChunk,
-            KaspadMessagePayload::RequestIbdBlocks(_) => KaspadMessagePayloadType::RequestIbdBlocks,
-            KaspadMessagePayload::UnexpectedPruningPoint(_) => KaspadMessagePayloadType::UnexpectedPruningPoint,
-            KaspadMessagePayload::IbdBlockLocator(_) => KaspadMessagePayloadType::IbdBlockLocator,
-            KaspadMessagePayload::IbdBlockLocatorHighestHash(_) => KaspadMessagePayloadType::IbdBlockLocatorHighestHash,
-            KaspadMessagePayload::RequestNextPruningPointUtxoSetChunk(_) => {
-                KaspadMessagePayloadType::RequestNextPruningPointUtxoSetChunk
-            }
-            KaspadMessagePayload::DonePruningPointUtxoSetChunks(_) => KaspadMessagePayloadType::DonePruningPointUtxoSetChunks,
-            KaspadMessagePayload::IbdBlockLocatorHighestHashNotFound(_) => {
-                KaspadMessagePayloadType::IbdBlockLocatorHighestHashNotFound
-            }
-            KaspadMessagePayload::BlockWithTrustedData(_) => KaspadMessagePayloadType::BlockWithTrustedData,
-            KaspadMessagePayload::DoneBlocksWithTrustedData(_) => KaspadMessagePayloadType::DoneBlocksWithTrustedData,
-            KaspadMessagePayload::RequestPruningPointAndItsAnticone(_) => KaspadMessagePayloadType::RequestPruningPointAndItsAnticone,
-            KaspadMessagePayload::BlockHeaders(_) => KaspadMessagePayloadType::BlockHeaders,
-            KaspadMessagePayload::RequestNextHeaders(_) => KaspadMessagePayloadType::RequestNextHeaders,
-            KaspadMessagePayload::DoneHeaders(_) => KaspadMessagePayloadType::DoneHeaders,
-            KaspadMessagePayload::RequestPruningPointUtxoSet(_) => KaspadMessagePayloadType::RequestPruningPointUtxoSet,
-            KaspadMessagePayload::RequestHeaders(_) => KaspadMessagePayloadType::RequestHeaders,
-            KaspadMessagePayload::RequestBlockLocator(_) => KaspadMessagePayloadType::RequestBlockLocator,
-            KaspadMessagePayload::PruningPoints(_) => KaspadMessagePayloadType::PruningPoints,
-            KaspadMessagePayload::RequestPruningPointProof(_) => KaspadMessagePayloadType::RequestPruningPointProof,
-            KaspadMessagePayload::PruningPointProof(_) => KaspadMessagePayloadType::PruningPointProof,
-            KaspadMessagePayload::Ready(_) => KaspadMessagePayloadType::Ready,
-            KaspadMessagePayload::BlockWithTrustedDataV4(_) => KaspadMessagePayloadType::BlockWithTrustedDataV4,
-            KaspadMessagePayload::TrustedData(_) => KaspadMessagePayloadType::TrustedData,
-            KaspadMessagePayload::RequestIbdChainBlockLocator(_) => KaspadMessagePayloadType::RequestIbdChainBlockLocator,
-            KaspadMessagePayload::IbdChainBlockLocator(_) => KaspadMessagePayloadType::IbdChainBlockLocator,
-            KaspadMessagePayload::RequestAnticone(_) => KaspadMessagePayloadType::RequestAnticone,
-            KaspadMessagePayload::RequestNextPruningPointAndItsAnticoneBlocks(_) => {
-                KaspadMessagePayloadType::RequestNextPruningPointAndItsAnticoneBlocks
-            }
-        };
-        res as u8
-    }
-
     fn incoming_flow_channel_size() -> usize {
         128
     }
@@ -487,8 +384,7 @@ impl Router {
         let (sender, receiver) = mpsc_channel(Self::incoming_flow_channel_size());
         let mut map = self.routing_map.write().await;
         for msg_type in msg_types {
-            let msg_id = msg_type as u8;
-            match map.insert(msg_id, sender.clone()) {
+            match map.insert(msg_type, sender.clone()) {
                 Some(_) => {
                     // Overrides an existing route -- panic
                     error!("P2P, Router::subscribe overrides an existing value: {:?}, router-id: {}", msg_type, self.identity);
@@ -508,9 +404,9 @@ impl Router {
             debug!("P2P, Route to flow got empty payload, router-id: {}", self.identity);
             return false;
         }
-        let key = Router::payload_to_u8(msg.payload.as_ref().unwrap());
+        let msg_type: KaspadMessagePayloadType = msg.payload.as_ref().unwrap().into();
         let map = self.routing_map.read().await;
-        if let Some(sender) = map.get(&key) {
+        if let Some(sender) = map.get(&msg_type) {
             sender.send(msg).await.is_ok()
         } else {
             false
